🚚 rename: DashboardContentView Ranking/ -> Rank/
